connect iteration
