+++ apps/studio/components/interfaces/ProjectHome/TopSection.tsx
@@ -1,5 +1,5 @@
 import { ActivityStats } from 'components/interfaces/ProjectHome/ActivityStats'
-import { ProjectConnectionHoverCard } from 'components/interfaces/ProjectHome/ProjectConnectionHoverCard'
+import { ProjectConnectionPopover } from 'components/interfaces/ProjectHome/ProjectConnectionPopover'
 import { ProjectPausedState } from 'components/layouts/ProjectLayout/PausedState/ProjectPausedState'
 import { ComputeBadgeWrapper } from 'components/ui/ComputeBadgeWrapper'
 import { InlineLink } from 'components/ui/InlineLink'
@@ -81,7 +81,7 @@ export const TopSection = () => {
                   />
                 </div>
               </div>
-              <ProjectConnectionHoverCard projectRef={project?.ref} />
+              <ProjectConnectionPopover projectRef={project?.ref} />
             </div>
           </div>
           <div className="mt-8">
