+++ apps/studio/components/layouts/Navigation/LayoutHeader/LayoutHeader.tsx
@@ -18,6 +18,7 @@ import { HelpButton } from 'components/ui/HelpPanel/HelpButton'
 import { getResourcesExceededLimitsOrg } from 'components/ui/OveragesBanner/OveragesBanner.utils'
 import { useOrgUsageQuery } from 'data/usage/org-usage-query'
 import { DevToolbarTrigger } from 'dev-tools'
+import dayjs from 'dayjs'
 import { AnimatePresence, motion } from 'framer-motion'
 import { useLocalStorageQuery } from 'hooks/misc/useLocalStorage'
 import { useSelectedOrganizationQuery } from 'hooks/misc/useSelectedOrganization'
@@ -59,7 +60,7 @@ const LayoutHeaderDivider = ({ className, ...props }: React.HTMLProps<HTMLSpanEl
 
 interface LayoutHeaderProps {
   customHeaderComponents?: ReactNode
-  breadcrumbs?: any[]
+  breadcrumbs?: unknown[]
   headerTitle?: string
   backToDashboardURL?: string
 }
@@ -97,6 +98,12 @@ export const LayoutHeader = ({
       return false
     }
   }, [orgUsage])
+
+  const connectButtonType =
+    selectedProject?.inserted_at &&
+    dayjs(selectedProject.inserted_at).isBefore(dayjs().subtract(5, 'day'))
+      ? 'default'
+      : 'secondary'
 
   // show org selection if we are on a project page or on a explicit org route
   const showOrgSelection = slug || (selectedOrganization && projectRef)
@@ -216,7 +223,7 @@ export const LayoutHeader = ({
                   }}
                 >
                   {IS_PLATFORM && gitlessBranching && <MergeRequestButton />}
-                  <ConnectButton />
+                  <ConnectButton buttonType={connectButtonType} />
                 </motion.div>
               )}
             </AnimatePresence>
